requirements.txt
docling
langchain
langchain_community
chromadb
ollama
python-dotenv
pydantic
tqdm
pytest
pytest-cov